+++ src/rex.mart/NEWS.rst
@@ -10,6 +10,8 @@ REX.MART Change Log
 
 - Added ability to specify additional HTSQL extensions to enable in the
   get_mart_db() and get_mart_etl_db() functions.
+- Added HTSQL connection caching to the web endpoint, as well as an associated
+  ``mart_htsql_cache_depth`` setting.
 
 
 0.1.0 (2016-01-29)